Refactored the histogram managers a bit
