+++ Jantimizer/Histograms/Managers/MySQLEquiDepthHistogramManager.cs
@@ -10,45 +10,14 @@ using Histograms.Models;
 
 namespace Histograms.Managers
 {
-    public class MySQLEquiDepthHistogramManager : IHistogramManager<IHistogram, IDbConnector>
+    public class MySQLEquiDepthHistogramManager : BaseEquiDepthHistogramManager
     {
-        public IDbConnector DbConnector { get; }
-        public List<IHistogram> Histograms { get; }
-        public List<string> Tables => Histograms.Select(x => x.TableName).Distinct().ToList();
-        public List<string> Attributes
-        {
-            get
-            {
-                var returnList = new List<string>();
-                foreach (var histogram in Histograms)
-                    returnList.Add($"{histogram.TableName}.{histogram.AttributeName}");
-                return returnList;
-            }
-        }
-        public int Depth { get; }
-
-        public MySQLEquiDepthHistogramManager(ConnectionProperties connectionProperties, int depth)
+        public MySQLEquiDepthHistogramManager(ConnectionProperties connectionProperties, int depth) : base(connectionProperties, depth)
         {
             DbConnector = new DatabaseConnector.Connectors.MySqlConnector(connectionProperties);
-            Histograms = new List<IHistogram>();
-            Depth = depth;
         }
 
-        public void ClearHistograms()
-        {
-            Histograms.Clear();
-        }
-
-        public void AddHistogram(IHistogram histogram)
-        {
-            if (string.IsNullOrWhiteSpace(histogram.TableName))
-                throw new ArgumentException("Table name cannot be empty!");
-            if (string.IsNullOrWhiteSpace(histogram.AttributeName))
-                throw new ArgumentException("Attribute name cannot be empty!");
-            Histograms.Add(histogram);
-        }
-
-        public async Task<List<Task>> AddHistogramsFromDB()
+        public override async Task<List<Task>> AddHistogramsFromDB()
         {
             ClearHistograms();
             List<Task> tasks = new List<Task>();
@@ -92,42 +61,6 @@ namespace Histograms.Managers
                 newHistogram.GenerateHistogram(attributeValues.Tables[0], attributeName);
                 Histograms.Add(newHistogram);
             }
-        }
-
-        public override string? ToString()
-        {
-            StringBuilder sb = new StringBuilder();
-            sb.AppendLine("Recorded Histograms:");
-            foreach (var histogram in Histograms)
-                sb.AppendLine(histogram.ToString());
-            return sb.ToString();
-        }
-
-        public IHistogram GetHistogram(string table, string attribute)
-        {
-            foreach (var gram in Histograms)
-                if (gram.TableName.Equals(table) && gram.AttributeName.Equals(attribute))
-                    return gram;
-
-            throw new ArgumentException("No histogram found");
-        }
-        public List<IHistogram> GetHistogramsByTable(string table)
-        {
-            List<IHistogram> grams = new List<IHistogram>();
-            foreach (var gram in Histograms)
-                if (gram.TableName.Equals(table))
-                    grams.Add(gram);
-
-            return grams;
-        }
-        public List<IHistogram> GetHistogramsByAttribute(string attribute)
-        {
-            List<IHistogram> grams = new List<IHistogram>();
-            foreach (var gram in Histograms)
-                if (gram.AttributeName.Equals(attribute))
-                    grams.Add(gram);
-
-            return grams;
         }
     }
 }
